--- a/Gemfile
+++ b/Gemfile
@@ -3,10 +3,8 @@ source 'https://rubygems.org'
 gem 'jekyll'
 gem 'pry'
 gem 'activesupport'
-gem 'uglifier'
 
 group :jekyll_plugins do
   gem 'jekyll-everypolitician'
-  gem 'jekyll-assets'
   gem 'jekyll-remote_csv', git: 'https://github.com/everypolitician/jekyll-remote_csv', branch: 'master'
 end