--- a/Gemfile
+++ b/Gemfile
@@ -2,7 +2,6 @@ source 'https://rubygems.org'
 
 gem 'jekyll'
 gem 'pry'
-gem 'activesupport'
 
 group :jekyll_plugins do
   gem 'jekyll-everypolitician'