--- a/_assets/javascripts/main.js
+++ b/_assets/javascripts/main.js
@@ -1,9 +1,0 @@
-//= require twitter_embed
-
-jQuery(function($) {
-  var $searchBox = $('.js-search-autocomplete-name')
-  $searchBox.select2();
-  $searchBox.change(function(e) {
-    window.location.href = $(this).val();
-  });
-});