--- a/_includes/head.html
+++ b/_includes/head.html
@@ -23,13 +23,13 @@
 
   <link href="http://www.kuvakazim.com/static/css/zimbabwe.f753013b63d6.css" rel="stylesheet" type="text/css" />
 
-  <link rel="icon" type="image/png" href="{% asset_path 'favicon.png' %}">
+  <link rel="icon" type="image/png" href="/images/favicon.png">
 
   <!-- select2 -->
   <link href="//cdnjs.cloudflare.com/ajax/libs/select2/4.0.1-rc.1/css/select2.min.css" rel="stylesheet" />
 
   <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
   <script src="//cdnjs.cloudflare.com/ajax/libs/select2/4.0.1-rc.1/js/select2.min.js"></script>
+  <script src="/javascripts/main.js"></script>
 
-  {% javascript 'main.js' %}
 </head>