--- a/_includes/header.html
+++ b/_includes/header.html
@@ -6,7 +6,7 @@
     <menu id="mobile-top-tools">
       <ul>
         <li><a href="#footer-menu" class="m-icon-nav">menu</a></li>
-        <li><a href="/search/" class="icon-search">{% image 'search-icon.png' %}</a></li>
+        <li><a href="/search/" class="icon-search"><img src="/images/search-icon.png"></a></li>
       </ul>
     </menu>