--- a/_layouts/areas.html
+++ b/_layouts/areas.html
@@ -10,7 +10,7 @@ layout: default
 </div>
 
 <div class="profile-pic">
-  {% image 'place-200x200.jpg' %}
+  <img src="/images/place-200x200.jpg">
 </div>
 
 <div class="profile-info">