--- a/_layouts/organizations.html
+++ b/_layouts/organizations.html
@@ -13,7 +13,7 @@ layout: default
   {% if organization.image.to_s == blank %}
     <img src="{{ organization.image }}" alt="{{ organization.name }}" width="50" />
   {% else %}
-    {% image 'organisation-210x210.jpg' %}
+    <img src="/images/organisation-210x210.jpg">
   {% endif %}
 </div>