--- a/javascripts/main.js
+++ b/javascripts/main.js
@@ -13,3 +13,11 @@ jQuery(function($) {
         }
     }(document, "script", "twitter-wjs");
 });
+
+jQuery(function($) {
+  var $searchBox = $('.js-search-autocomplete-name')
+  $searchBox.select2();
+  $searchBox.change(function(e) {
+    window.location.href = $(this).val();
+  });
+});