… Gemfile.lock
… images/favicon.png
… images/organisation-210x210.jpg
… images/person-210x210.jpg
… images/place-200x200.jpg
… images/search-icon.png