This speeds up the builds from ~25s to ~13s. This is most of the low hanging fruit I think. Amazingly removing [this line in `header.html`](https://github.com/mysociety/kuvakazim/blob/7fd2f85e058aff8ea645f8c83235efda4bed0fbf/_includes/header.html#L14) reduces the build time from ~13s to ~6.5s! So it seems that `for` loops in liquid are surprisingly expensive. I've opened a ticket on Liquid (https://github.com/Shopify/liquid/issues/678) to see if there's some thing obvious we're missing, but for now we might have to look at implementing the search in a way that doesn't use so many for loops.